Hide IConnectionDetailsInternal and IDeltaHandlerStrategy and remove unused events
