+++ packages/common/container-definitions/src/deltas.ts
@@ -5,18 +5,16 @@
 
 import {
 	IDisposable,
+	IErrorEvent,
 	IEventProvider,
 	IEvent,
-	IErrorEvent,
 } from "@fluidframework/common-definitions";
 import { IAnyDriverError } from "@fluidframework/driver-definitions";
 import {
-	ConnectionMode,
 	IClientConfiguration,
 	IClientDetails,
 	IDocumentMessage,
 	ISequencedDocumentMessage,
-	ISignalClient,
 	ISignalMessage,
 	ITokenClaims,
 } from "@fluidframework/protocol-definitions";
@@ -40,47 +38,6 @@ export interface IConnectionDetails {
 	 * that is likely to be more up-to-date.
 	 */
 	checkpointSequenceNumber: number | undefined;
-}
-
-/**
- * Internal version of IConnectionDetails with props are only exposed internally
- * @deprecated 2.0.0-internal.5.2.0 - Intended for internal use only and will be removed in an upcoming relase.
- */
-export interface IConnectionDetailsInternal extends IConnectionDetails {
-	/**
-	 * @deprecated 2.0.0-internal.5.2.0 - Intended for internal use only and will be removed in an upcoming relase.
-	 */
-	mode: ConnectionMode;
-	/**
-	 * @deprecated 2.0.0-internal.5.2.0 - Intended for internal use only and will be removed in an upcoming relase.
-	 */
-	version: string;
-	/**
-	 * @deprecated 2.0.0-internal.5.2.0 - Intended for internal use only and will be removed in an upcoming relase.
-	 */
-	initialClients: ISignalClient[];
-	/**
-	 * @deprecated 2.0.0-internal.5.2.0 - Intended for internal use only and will be removed in an upcoming relase.
-	 */
-	reason: string;
-}
-
-/**
- * Interface used to define a strategy for handling incoming delta messages
- * @deprecated 2.0.0-internal.5.2.0 - Intended for internal use only and will be removed in an upcoming relase.
- */
-export interface IDeltaHandlerStrategy {
-	/**
-	 * Processes the message.
-	 * @deprecated 2.0.0-internal.5.2.0 - Intended for internal use only and will be removed in an upcoming relase.
-	 */
-	process: (message: ISequencedDocumentMessage) => void;
-
-	/**
-	 * Processes the signal.
-	 * @deprecated 2.0.0-internal.5.2.0 - Intended for internal use only and will be removed in an upcoming relase.
-	 */
-	processSignal: (message: ISignalMessage) => void;
 }
 
 /**
@@ -127,17 +84,7 @@ export interface IDeltaManagerEvents extends IEvent {
 	/**
 	 * @deprecated No replacement API recommended.
 	 */
-	(event: "allSentOpsAckd", listener: () => void);
-
-	/**
-	 * @deprecated No replacement API recommended.
-	 */
 	(event: "pong", listener: (latency: number) => void);
-
-	/**
-	 * @deprecated No replacement API recommended.
-	 */
-	(event: "processTime", listener: (latency: number) => void);
 
 	/**
 	 * Emitted when the {@link IDeltaManager} completes connecting to the Fluid service.
