+++ packages/common/container-definitions/src/index.ts
@@ -17,13 +17,11 @@ export {
 } from "./browserPackage";
 export {
 	IConnectionDetails,
-	IConnectionDetailsInternal,
-	IDeltaHandlerStrategy,
 	IDeltaManager,
 	IDeltaManagerEvents,
-	IDeltaSender,
 	IDeltaQueue,
 	IDeltaQueueEvents,
+	IDeltaSender,
 	ReadOnlyInfo,
 } from "./deltas";
 export {
