+++ packages/common/container-definitions/src/test/types/validateContainerDefinitionsPrevious.generated.ts
@@ -352,26 +352,14 @@ use_old_InterfaceDeclaration_IConnectionDetails(
 /*
 * Validate forward compat by using old type in place of current type
 * If breaking change required, add in package.json under typeValidation.broken:
-* "InterfaceDeclaration_IConnectionDetailsInternal": {"forwardCompat": false}
+* "RemovedInterfaceDeclaration_IConnectionDetailsInternal": {"forwardCompat": false}
 */
-declare function get_old_InterfaceDeclaration_IConnectionDetailsInternal():
-    TypeOnly<old.IConnectionDetailsInternal>;
-declare function use_current_InterfaceDeclaration_IConnectionDetailsInternal(
-    use: TypeOnly<current.IConnectionDetailsInternal>);
-use_current_InterfaceDeclaration_IConnectionDetailsInternal(
-    get_old_InterfaceDeclaration_IConnectionDetailsInternal());
 
 /*
 * Validate back compat by using current type in place of old type
 * If breaking change required, add in package.json under typeValidation.broken:
-* "InterfaceDeclaration_IConnectionDetailsInternal": {"backCompat": false}
+* "RemovedInterfaceDeclaration_IConnectionDetailsInternal": {"backCompat": false}
 */
-declare function get_current_InterfaceDeclaration_IConnectionDetailsInternal():
-    TypeOnly<current.IConnectionDetailsInternal>;
-declare function use_old_InterfaceDeclaration_IConnectionDetailsInternal(
-    use: TypeOnly<old.IConnectionDetailsInternal>);
-use_old_InterfaceDeclaration_IConnectionDetailsInternal(
-    get_current_InterfaceDeclaration_IConnectionDetailsInternal());
 
 /*
 * Validate forward compat by using old type in place of current type
@@ -497,26 +485,14 @@ use_old_TypeAliasDeclaration_ICriticalContainerError(
 /*
 * Validate forward compat by using old type in place of current type
 * If breaking change required, add in package.json under typeValidation.broken:
-* "InterfaceDeclaration_IDeltaHandlerStrategy": {"forwardCompat": false}
+* "RemovedInterfaceDeclaration_IDeltaHandlerStrategy": {"forwardCompat": false}
 */
-declare function get_old_InterfaceDeclaration_IDeltaHandlerStrategy():
-    TypeOnly<old.IDeltaHandlerStrategy>;
-declare function use_current_InterfaceDeclaration_IDeltaHandlerStrategy(
-    use: TypeOnly<current.IDeltaHandlerStrategy>);
-use_current_InterfaceDeclaration_IDeltaHandlerStrategy(
-    get_old_InterfaceDeclaration_IDeltaHandlerStrategy());
 
 /*
 * Validate back compat by using current type in place of old type
 * If breaking change required, add in package.json under typeValidation.broken:
-* "InterfaceDeclaration_IDeltaHandlerStrategy": {"backCompat": false}
+* "RemovedInterfaceDeclaration_IDeltaHandlerStrategy": {"backCompat": false}
 */
-declare function get_current_InterfaceDeclaration_IDeltaHandlerStrategy():
-    TypeOnly<current.IDeltaHandlerStrategy>;
-declare function use_old_InterfaceDeclaration_IDeltaHandlerStrategy(
-    use: TypeOnly<old.IDeltaHandlerStrategy>);
-use_old_InterfaceDeclaration_IDeltaHandlerStrategy(
-    get_current_InterfaceDeclaration_IDeltaHandlerStrategy());
 
 /*
 * Validate forward compat by using old type in place of current type
